refactor: replace web component
There's a bug in Chromium that makes text input bug out.
