+++ packages/client/public/headers/headers.css
@@ -20,3 +20,32 @@
 	flex-direction: column;
 	gap: var(--spacing-24);
 }
+
+.container {
+	display: flex;
+	flex-direction: column;
+	gap: var(--spacing-8);
+}
+
+.input {
+	display: block;
+	min-width: 280px;
+	width: 100%;
+}
+
+.input-checkbox {
+	display: flex;
+	align-items: center;
+}
+
+.input-checkbox input {
+	cursor: pointer;
+	accent-color: var(--button-primary-bg-color);
+}
+
+.actions {
+	display: flex;
+	flex-direction: row;
+	justify-content: flex-end;
+	gap: var(--spacing-16);
+}
