+++ packages/client/public/headers/headers.html
@@ -24,6 +24,8 @@
 					<option value="">---</option>
 					<option value="desktop">Desktop</option>
 					<option value="mobile">Mobile</option>
+					<option value="hybrid-ios">Hybrid (iOS)</option>
+					<option value="hybrid-android">Hybrid (Android)</option>
 				</select>
 			</div>
 			<form id="headers" class="headers-form">
